Create a single data frame from a vector of Twitter handles

Create a single data frame from a vector of Twitter handles

qf_bind_rows_tweets(
  users = NULL,
  lists = NULL,
  since = Sys.Date() - 31,
  include_rts = FALSE,
  add_date_column = TRUE,
  save = TRUE,
  twitter_token = NULL
)

Arguments

users
A character vector of handles or id of Twitter users.

lists
A character vector of twitter lists, either as IDs or in the form `user/slug``

since
A date, expressed in the form Y-M-D (e.g. "2018-12-31")

save
Logical, defaults to TRUE. If TRUE, the merged data frame is stored in the following location: `tweets_all/tweets.all.rds`

Examples