Fin the ID of Twitter lists.

Fin the ID of Twitter lists.

qf_find_list_id(slug, owner_user, cache_lists = TRUE,
  overwrite_lists = FALSE, twitter_token = NULL)

Arguments

cache_lists
Logical, defaults to TRUE. If TRUE, stores locally the lists owned by a given user, in a subfolder `lists_by_user`. If list has already been downloaded, it just loads it. To overwrite, set `overwrite` to TRUE.

overwrite_lists
Logical, defaults to FALSE. If TRUE, it overwrites previously downloaded lists.

twitter_token
A twitter token generated by rtweet.

n
An integer, number of new tweets to request per users, defaults to 200.

Examples